+++ app/src/main/java/com/maurodev/slottdd/data/FirstReel.kt
@@ -1,5 +1,0 @@
-package com.maurodev.slottdd.data
-
-data class FirstReel( // Aun no sabemos si lo necesito
-    val value: SlotValue
-)