+++ app/src/main/java/com/maurodev/slottdd/data/slot/FirstReel.kt
@@ -1,0 +1,5 @@
+package com.maurodev.slottdd.data.slot
+
+data class FirstReel( // Aun no se si lo necesito
+    val value: SlotValue
+)