+++ app/src/main/java/com/maurodev/slottdd/domain/slot/SlotManager.kt
@@ -1,4 +1,4 @@
-package com.maurodev.slottdd.domain
+package com.maurodev.slottdd.domain.slot
 
 /**
  * Esta clase se encarga de manejar la logica e implementacion del slot.